test: resolve duplications
