+++ src/tests/utils/expectTitlePost.ts
@@ -1,0 +1,5 @@
+import { screen } from '@testing-library/react';
+
+const editPost = 'Editar um post';
+
+export const expectTitlePost = () => expect(screen.getByRole('heading', { name: editPost })).toBeInTheDocument();
